true_color_fix added

+++ defense_of_the_future/video_gr.h
@@ -18,7 +18,7 @@
  * @param mode 16-bit VBE mode to set
  * @return Virtual address VRAM was mapped to. NULL, upon failure.
  */
-void *vg_init(unsigned short mode);
+void *vg_init(unsigned long mode);
 
 /**
  * @brief Fills the screen with the input color
